server/requirements.txt
requests
fastapi
uvicorn
reportlab
loguru
logtail-python